requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==36.0.1
defusedxml==0.7.1
distlib==0.3.3
dj-database-url==0.5.0
Django==3.2.8
django-allauth==0.47.0
django-cors-headers==3.10.0
django-cors-middleware==1.5.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
filelock==3.3.0
gunicorn==20.1.0
idna==3.3
mysqlclient==2.1.0
oauthlib==3.1.1
Pillow==8.4.0
platformdirs==2.4.0
psycopg2-binary==2.9.1
pycparser==2.21
PyJWT==2.3.0
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
virtualenv==20.8.1
waitress==2.0.0
whitenoise==5.3.0
